activitylogのファイルをさらに分割
テスト側も合わせて分割。
