+++ src/routes/api_v1.js
@@ -11,7 +11,7 @@ var router = express.Router();
 var console_output = require("../api/debugger.js").console_output;
 var lib = require("../api/factory4require.js");
 var factoryImpl = { // require()を使う代わりに、new Factory() する。
-    "sql_lite_db" : new lib.Factory4Require("../api/activitylog/activitylog.js")
+    "sql_activitylog" : new lib.Factory4Require("../api/activitylog/index.js")
 };
 
 
@@ -80,7 +80,7 @@ var responseAnomaly = function( res, err ){
 
 
 router.post("/setup1st", function(req, res, next){
-	var api_vi_activitylog_setup = factoryImpl.sql_lite_db.getInstance().api_vi_activitylog_setup;
+	var api_vi_activitylog_setup = factoryImpl.sql_activitylog.getInstance().api_vi_activitylog_setup;
 	var dataPost = req.body; // app.jsで「app.use(bodyParser.json());」してるので、bodyプロパティが使える。
 
 	return api_vi_activitylog_setup( null, dataPost ).then((result)=>{
@@ -91,7 +91,7 @@ router.post("/setup1st", function(req, res, next){
 });
 
 router.post("/signup", function(req, res, next){
-	var api_vi_activitylog_signup = factoryImpl.sql_lite_db.getInstance().api_vi_activitylog_signup;
+	var api_vi_activitylog_signup = factoryImpl.sql_activitylog.getInstance().api_vi_activitylog_signup;
 	var dataPost = req.body;
 	return api_vi_activitylog_signup( null, dataPost ).then((result)=>{
 		responseNormal( res, result );
@@ -101,7 +101,7 @@ router.post("/signup", function(req, res, next){
 });
 
 router.post("/remove", function(req, res, next){
-	var api_vi_activitylog_remove = factoryImpl.sql_lite_db.getInstance().api_vi_activitylog_remove;
+	var api_vi_activitylog_remove = factoryImpl.sql_activitylog.getInstance().api_vi_activitylog_remove;
 	var dataPost = req.body;
 	return api_vi_activitylog_remove( null, dataPost ).then((result)=>{
 		responseNormal( res, result );
@@ -113,7 +113,7 @@ router.post("/remove", function(req, res, next){
 
 
 router.get('/show', function(req, res, next) {
-	var api_v1_activitylog_show = factoryImpl.sql_lite_db.getInstance().api_v1_activitylog_show;
+	var api_v1_activitylog_show = factoryImpl.sql_activitylog.getInstance().api_v1_activitylog_show;
 
 	return api_v1_activitylog_show(req.query, null ).then((result)=>{
 		responseNormal( res, result );
@@ -123,7 +123,7 @@ router.get('/show', function(req, res, next) {
 });
 
 router.post("/add", function(req, res, next){
-	var api_v1_activitylog_add = factoryImpl.sql_lite_db.getInstance().api_v1_activitylog_add;
+	var api_v1_activitylog_add = factoryImpl.sql_activitylog.getInstance().api_v1_activitylog_add;
 	var dataPost = req.body;
 	
 	return api_v1_activitylog_add( null, dataPost ).then((result)=>{
@@ -134,7 +134,7 @@ router.post("/add", function(req, res, next){
 });
 
 router.post("/delete", function(req, res, nest) {
-	var api_v1_activitylog_delete = factoryImpl.sql_lite_db.getInstance().api_v1_activitylog_delete;
+	var api_v1_activitylog_delete = factoryImpl.sql_activitylog.getInstance().api_v1_activitylog_delete;
 	var dataPost = req.body;
 	
 	return api_v1_activitylog_delete( null, dataPost).then((result)=>{
